// src/model/Interface/applicationform.module.ts
export interface Studentclass {
    studentEname: string;
    studentHname: string;
    fatherEname: string;
    fatherHEname: string;
    motherEname: string;
    motherHname: string;
    caste: string;
    gender: string;
    dob: Date;
    dobwords: string;
    mobile: string;
    adhivash: string;
    presentAddress: string;
    presentpincode: string;
    permanentAddress: string;
    permanentpincode: string;
    standard: string;
    stream: string;
    studycenterzila: string;
    studycenter: string;

  }